implemnted choosing trigger in GUI
this ugly commit successfully implements the ability to choose between the 'lock-screen' and 'soft-shutdown' trigger in the GUI's Settings Screen.

For a long time I thought that the JSON file that the Settings widget used was whre it saved the settings. That's totally wrong. It's actually just a *template* that tells the Settings widget what options are possible. It doesn't store the *values* that the user selects to the json file. It always stores *that* to the .ini file.

So this commit creats the json file in the src/packages/buskill/ directory. It complained that the default value was't already set, so I had to also add a cooresponding default value to build_config() for the 'trigger' key. It works!

Of course, this commit only adds the GUI selector and its persistance to the config file.

TODO: cleanup the code

TODO: actually check the config option when arming/disarming. right now it isn't actually used at all.