finally able to remove the 'buskill' title
This commit is the fruitful effort of tying to figure out the structure of the Kivy Settings classes. Finally, this commit, I was able to remove the title of the Settings panel so it no longer says 'buskill' at the top -- not needed since we only have one "section"

  s.children[0].children[0].children[0].remove_widget( s.children[0].children[0].children[0].children[1] )

This is pretty ridiculious, but I learend:

 1. the 'children' attribute of a given kivy widget is not recursive, so you have to use walk()
 2. If you run remove_widget() directly on a root's grandparent, it won't work. You need to find the right widget and remove it directly from its parent
 3. Somehow the title of the section appears before the options inside that title?

This commit also overrides three Settings classes from kivy/uix/setings.py, but I'm still experiementing. My goal was to just change the way things are displayed, but it's buried in so many levels of inheretance that it's not trival.

TODO: figure out the structure better and document it here:

 * https://github.com/BusKill/buskill-app/issues/16