finally fixed overriding the SettingsPanel in kv-language
This commit finally solved the puzzling issue where I kept trying to change the layout for a given widget that I'm overriding (in this case the SettingsPanel with BusKillSettingsPanel), but it somehow just added the layout on-top of the existing layout. For example, I couldn't get rid of the label at the top of SettingsPanel. Even if it was absent in BusKillSettingsPanel and I used only BusKillSettingsPanel.

Solution: If you want to completely override a given classes' layout in kv-language, you must prepend a dash (-) before the definition.

For example, don't use this:

  <BusKillSettingsPanel>:
      spacing: 5
      padding: 5
      size_hint_y: None
      height: self.minimum_height

Use this:

  <-BusKillSettingsPanel>:
      spacing: 5
      padding: 5
      size_hint_y: None
      height: self.minimum_height

See also:

 * https://stackoverflow.com/questions/74467933/how-to-override-built-in-kivy-widgets-and-their-kv-language-layouts
 * https://kivy.org/doc/stable/api-kivy.lang.html#redefining-a-widget-s-style
 * https://github.com/BusKill/buskill-app/issues/16#issuecomment-1317832869