half-fix of double-popup issue
for some reason I've been fighting with an issue where the overloaded SettingsItem row, when clicked, produces two pop-ups. You can't see them on-top of each other, but two things happen:

1. The somewhat-opaque view behind the Popup is darker than normal -- because it's darkened twice
2. You have to click twice to escape both popups

I can't figure out why two pop-ups come. I tried to print a ton of info about both popups, but I couldn't get any context. One thing I did learn, when I diff'd them, there's 4x distinctson: the pointer to the popup itself is unique (so they are, in-fact, two distinct object instances), and the _trigger_layout, canvas, and _proxy_ref are also all distinct

I found that _create_popup() is called twice for some reason, and that on the first call self.popup is None and on the second call, it's already defined.

So this commit half-fixes the issue by simply checking if self.popup is already defined. If it is, it just does a 'return'.

It's only half-fixed, however, because once the popup is closed you can't re-open it again by tapping the SettingItem row