fixed overriding the kv-language of the SettingsPanel to get rid of the title 'buskill' label
this commit adds-back the definition of the overloaded create_json_panel() method in the BusKillSettings class that overloads kivy.uix.settings.Settings.

This is necessary because this is the method that defines the 'panel' variable as an instantiation of the BusKillSettingsPanel() object. If it's not overloaded, then the 'panel' instance field will become a SettingsPanel() object