got rid of BusKillSettingsPanel
This commit un-does the last commit and simplifies things by simply overriding the SettingsPanel in the kv-langauge (as opposed to creating a new BusKilSettingsPanel that inherets it). This works since we don't ever plan to have more than one type of SettingsPanel, so we don't need to actually preserve the default functioanlity/style. And it makes our code simpler.