removed 'content' of SettingItem
This commit commented-out the "content" section of the BusKillSettingItem style in kv-language.

This 'content' BoxLayout would appear to the right-hand side of the setting, to display what the 'value' of the current setting's 'key' is.

By contrast, In Material Design, the value of the setting would appear under the setting's 'key'.

By removing this, we make the UX more familiar to the users by complying with the Material Design spec, and we also leave more horizontal space for a new icon to appear to the left of the setting, which is also good for UX