placholder if no icon defined for SettingItem
This commit is a hack that inserts a dummy "space" icon with 0% opacity (alpha '00') if no icon was defined for a given SettingsItem.

Note that there is no "space" icon, and I didn't want to guess as to how wide a material design icon should be in pixels.

So this hack makes sure that there is a gap of equal width in a row where no icon is defined.