added route to get screen manager from widgets
This commit implements a recursive function get_screen_manager() that, given some kivy widget object, will iterate up the tree (checking the widget's parent attribute) until it finally reaches the Screen, which has the 'manager' attribute, which is our root Screen Manager.

This allows us to be able to dynamically create new screens and switch to them from within a widget in our Settings Panel.

This is important because the Popup() is too limited and shitty. It's better if we stick to the Material Design spec and, every time a user wants to change an "option" setting, we just open a new screen with a scrollable view for each option. Like here:

 * https://m2.material.io/design/platform-guidance/android-settings.html#label-and-secondary-text