added confirmation when choosing 'soft-shutdown'
This commit adds logic that presents the user with a 'confirmation' and makes them confirm before choosing an option where a non-empty-string is defined in the cooresponding element in the "confirmation" list of the settings json.

For example, our 'lock-screen' option has no confirmation (an empty string). But our 'soft-shutdown' (and 'hard-shutdown' example) both require the user to confirm the risk of data loss and data corruption before actually changing the setting, respectfully.

TOOD: actually propogate these settings down to the .ini file (see how it was done by the popup in the actual built-in kivy Settings widget)