actually got GUI changes to settings to write to .ini file
this commit swaps the instance field of the OptionItem (one of many options available for a ComplexOption) for 'value' to 'parent_option'. I probably should come up with a better var name, but this means that it's the "parent" of the "OptionItem" -- that is the relevant ComplexOption that's parent to a bunch of OptionItems. This is distinct from a "parent", which is a kivy-built-in and refers to its place in the widget hierarchy.

By using 'parent_option', we can reference the whole ComplexOption's instance fields (like key & value) rather than store it twice for each OptionItem.

So now we use "Config" to write() the new 'parent_option.value' to 'parent_option.key' in the 'buskill' section of the .ini file (that's already defined)

It works!