* added "with 'X' trigger" message when armed  * added debug logging when changing the trigger
So we're currently in an awkward situation: when a user changes the trigger when BusKill is armed, what should we do? Because some triggers might require root (eg soft-shutdown on MacOS), we actually need to set the bk instance's trigger instance field and then disarm and re-arm. But we should probably ask the user if they want to do that.

Anyway, now that we support multiple triggers, I decided that we really need to make it clear on the main "BusKill is armed" message which trigger it's currently armed with.

And now we see that my logic in this branch thus-far doesn't do shit to change the trigger until the app is rebooted.

TODO: update the bk instance's 'trigger' instance field when we write to the Config file.