changed cinnamon screensaver lock from fallback to asynch call
I discovered that on Cinnamon Desktop Environment (eg Linux Mint Cinnamon), the `xdg-screensaver` is actually present, and when you call it to lock the screen, it exits 0 and does not output any errors. But, in fact, it doesn't lock the screen at all.

Therefore, this commit changes the logic from "fallback to the cinnamon lockscreen command" to "just attempt it at the start" because we cannot know if the first command failed on Cinnamon due to what I'll call an upstream bug (either the command shouldn't exist or it should fail non-zero or with an error message of some kind)

 * https://github.com/BusKill/buskill-app/issues/64