added some other types for testing
 * https://kivy.org/doc/stable/api-kivy.uix.settings.html#create-a-panel-from-json

As I cleanup and remove chunks of code that probably aren't necessary, I want to make sure that I don't break functionality with the other built-in kivy types. Currently we're only using my custom "ComplexOption", but I may want to use some of these in the future.

For example, to set auxiliary triggers, I may want to use the "path" type, which allows the user to pick a file on their computer (eg a script) to execute when the trigger executes

 * https://github.com/BusKill/buskill-app/issues/62

So far I tested all of these and they seem to work. There was an additional "Color" picker type, but that failed with an error. Maybe I have too old a version of kivy, but I'm content if our app doesn't allow users to pick colors for now..