removed all test options
In this commit I removed all the test options, which included many of the kivy built-in SettingItem types as well as our 'hard-shutdown', which is not a trigger that we've actually implemented yet.

At this point I think we may have finished with this feature.

TODO: test that the triggers actually work on all three platforms. If all is well, then merge to dev.