move the call of set_trigger() to the togggle() function
This commit fixes an issue where initalizing the BusKill object already calls set_trigger() with the default setting. This is an issue when a user runs `buskill --trigger lock-screen ...` if the `trigger` is set to `soft-shudown` in the config file.

In ^ that case, the user would be prompted for their password to launch a root_child, even though they don't need one for the explicitly-specified lock-screen trigger

By moving it to toggle(), we can avoid such a situation