have usb_handler tell parent to execute TRIGGER_FUNCTION()
This commit reworks the capturing of usb events such that the child process usb_handler tells the parent (via a Queue) to execute the TRIGGER_FUNCTION, as opposed to the child calling the TRIGGER_FUNCTION() itself.

This is important because some objects can't be pickled down to the child usb_handler process. While it works most of the time, this became an issue when the usb_handler needed to send the 'soft-shutdown' command to the root_child process, but that needed to happen via a ctypes pointer to the IO of the root_child process, which couldn't be pickled. As such, the usb_handler child process couldn't send messages to the root_child child process. The error was this:

  DEBUG: Attempting to send 'soft-shutdown' command to root child
  ERROR: Failed to send 'soft-shutdown' command to root child
  	'BusKill' object has no attribute 'root_child'

For more information, see:

 * https://github.com/BusKill/buskill-app/issues/73#issuecomment-1592414968
 * https://stackoverflow.com/questions/76486644/child-process-call-function-in-parent-process-python-multiprocessing?noredirect=1#comment134862811_76486644