permit root_child spawning if owned by 'admin' group
This commit also allows BusKill to launch the root_child process if it's owned by the 'admin' group (with gid=80). It appears that on some installs it gets owned by 'staff' and on others it gets owned by 'admin'. All users are already members of 'staff', so that's not a problem. But 'admin' is an even-more restrictive group, so it's actually safer to permit execution of a script as root that's owned by 'admin' than one that's owned by 'staff'

This should help fix bug #77

 * https://github.com/BusKill/buskill-app/issues/77