added "font" to Settings
this commit tests adding a very simple options-type item to the Settings Menu with two bullshit settings to choose-from.

I think we can get away with just using an "options" type setting for this one, but it is a bit complicated in that the contents needs to be dynamically generated based on what fonts we find on the system.

I'm not sure yet how to inject options into the setting. I'll work on that in the next commit