POC of font as a ComplexOption
I hopped up & down the SettingsPanel but I couldn't actually find the list of options that came-in from the json file so that I could modify what's viewed in the app at runtime.

Instead, I changed it to a ComplexOption and very quickly was able to hack this commit-up, which adds a hard-coded option "some_font". Already it's working in the GUi and writing to the config file.

TODO: edit it such that it actually adds all the fonts