POC: dynamically adding all fonts to ComplexOption items
Good news: I did succed in being able to populate the ComplexOption with a list of OptionItems for every font found on the system.

Bad news: It crashed my system.

Most of these fonts are probably useless anyway. For example, I randomly picked-one and it was a font for the Tamil language. We need a way to par-down the list to be shorter.

TODO:
 1. remove symlinks (duplicates) from the list of fonts
 2. remove fonts that don't have glyphs for latin characters

For #2, I found some code on SE that uses the python module fonttools, but I actually liked this one better -- as it just uses the built-in module unicodedata

 * https://superuser.com/questions/876572/how-do-i-find-out-which-font-contains-a-certain-special-character

For more info, see:

 * https://github.com/BusKill/buskill-app/issues/55#issuecomment-2002303359