attempting to use RecycleView
This is not working. The biggest issue that I'm having with getting RecycleView working is that I used positional arguments for the creation of BusKillOptionItem()

My understanding is that RecycleView maintains a list of some type of widget. In my case, I want RecycleView to be a set of my custom widget = BusKillOptionItem()s.

But the way that RecycleView creates & recycles these objects is by a dictionary, which gets passed into the objects when they're created as kwargs.

So in this commit I tried changing from positional args to kwargs, but it doesn't appear to be working yet. For more info, see:

 * https://github.com/BusKill/buskill-app/issues/55#issuecomment-2002321283
 * https://groups.google.com/g/kivy-users/c/TLZoBXX5zZs
 * https://kivy.org/doc/stable/api-kivy.uix.recycleview.html