I fixed the BusKillOptionItem()!!!!!!
This commit creates a on_manager() function, which is called by Kivy whenever the Kivy Property for the object named "manager" gets a new value.

 * https://stackoverflow.com/questions/78177697/how-to-access-kivy-properties-from-within-its-self-on-init/78177807#78177807
 * https://kivy.org/doc/stable/api-kivy.properties.html#observe-using-on-propname

This finally allows us (well, actually the RecycleView) to setup the instances of BusKillOptionItem() while still having access to the manager, main screen, and dialog objects that were originally passed to it as positional arguments, but are now Kivy Properites.