fixed scroll erroneously being detected as a click on Settings
For some reason when someone was simply scrolling through all the options in a ComplexSetting screen and they reached the end, kivy would pass that scroll event as a click event to the OptionItem -- making us change the Config as if they had clicked on it.

I don't know exactly why kivy does this, but it probably has to do with muti-touch gestures. Kivy thinks its better to bind and call all visible widgets (not just the ones touched) because gestures can sometimes end not where they started. This is different, but maybe related:

 * https://kivy.org/doc/stable/guide/inputs.html#touch-event-basics

> By default, touch events are dispatched to all currently displayed widgets. This means widgets receive the touch event whether it occurs within their physical area or not.
>
> This can be counter intuitive if you have experience with other GUI toolkits. These typically divide the screen into geometric areas and only dispatch touch or mouse events to the widget if the coordinate lies within the widgets area.
>
> This requirement becomes very restrictive when working with touch input. Swipes, pinches and long presses may well originate from outside of the widget that wants to know about them and react to them.

The solution to this was to check the touch.button instance field, which was either "left" or one of "scrollup", "scrolldown", "scrollright", and "scrollleft"

 * https://kivy.org/doc/stable/api-kivy.uix.behaviors.compoundselection.html#kivy.uix.behaviors.compoundselection.CompoundSelectionBehavior.goto_node

For more info, see:

 * https://stackoverflow.com/questions/78183125/scrolling-causes-click-on-touch-up-event-on-widgets-in-kivy-recycleview/78183647#78183647
 * https://github.com/BusKill/buskill-app/issues/55#issuecomment-2005131382