debugging radio button mismatch
for some reason the "lock-screen" radio button icon is being set to the "unchecked" radio button icon when the screen is first loaded, even though 'lock-screen' is the value of the 'trigger' in the buskill.ini config file

If I click on 'lock-screen' ItemOption then it says it's already set, so it clearly knows this.

After a few hours of debugging, I think I'm hitting some sort of race condition with the on_manager() and on_parent_option() functions. I can't seem to find any sort of on_load() or on_start() built-in function for kivy that's called once *all* of the properties of the object are initialized.

TODO:

1. try to use the on_<propername>() functions to work together as a wrapper to call some init2() funtion. If not all of the properties are defined, just return. If all are defined, then the final on_prop() function to be called will actually call the init2() function

2. try to update the rv data directly with refresh_from_data()