fixing the inefficient data loop and radio button on-load bug
this commit just reverts to an old version of the buskill.kv & buskill_gui.py files, with only the default_size line changed in buskill.kv.

I confirmed that all of the radio buttons now get updated as-desired on-launch.

I also confirmed that we don't need that stupid loop that updates all of the data points a few hundred times on every init() (executed a few hundred times)

The 'manager' stuff crept back. I could try to fix that again; I'd be surprised if that was the culprit.