added human-readable name for option to BusKillOptionItem
This commit adds a new instance field 'option_human' to the BusKillOptionItem object. Previously, what would be displayed to the user for the option was the "vale" -- this became an issue with the fonts because what we want to save to the cbuskill.ini Config is the absolute path to the font -- but that's too damn long to display in the GUi as the name for the option.

So this new "option_human" is a shorter human-readable string that will be displayed in the GUI for the option. In this case, it's just the basename() of the font