removed references to "roboto" font face in all text markup
This commit eliminates specifiably referencing the "roboto" font everywhere in the kivy text markup and where we update label's .text properties.

Currently we only have two font types that are explicitly defined:

 1. bkmono
 2. mdicons

The "bkmono" can be changed as-needed, but currently it's set on startup either to "RobotoMono" if the user hasn't set anything or whatever the user set their font-to. Note this means if the user sets their own font, then there will be no mono fonts displayed (unless they happened to set this to a mono font). This is also the case for bold and italic (in this commit I also replaced references to RobotoMedium with the more-robust "bold" attribute).

In the future we could consider writing some logic that tries to figure-out the Bold and Italic and BoldItalic font faces for the user's selected font, but I don't think that's so critical right now.

Currently there's also an issue where the user has to restart twice to get the font face to apply.

TODO: when the user changes their font face in Settings, update the kivy 'default_font' and then update all labels in the app to use the updated font (without requiring restarts)