added "value_human" to BusKillSettingComplexOptions and replaces 'gui_font_face' with 'default_font'
This commit removes the 'gui_font_face' setting from the buskill.ini file. Instead I decided to just use the existing kivy 'default_font'. The major difference between these two was that 'gui_font_face' was just the path to the font file whereas 'default_fonts' is a list where the first element is the human-readable name of the font and the following 4 elements are the regular, bold, italic, and bolditalic files for the font face.

I realized that we needed a better way to store the human-readable font name in the value itself, so just simplifying the Config to use the 'default_font' setting and its existing format works.

This commit updates the BusKillSettingComplexOptions object to have a 'value_human' instance field. When the object is instantiated, it checks to see if the 'value' is actually a list (it's a string and we just use the json module to check if its a string of a list). If it is a list, then we set "BusKillSettingComplexOptions.value_human" from None to the first element of the list (for 'default_font', that's the human-readable font name). If the value is not a string of a list, then it behaves as before.