hardcoding the default_font :(
This commit hardcodes the default_font to Roboto.

Actually all I'm doing here is copying these lines from kivy/config.py into our BusKillApp.build_config() function, but I don't like it because if Kivy changes from Roboto to something else in the future, this will break.

 * https://github.com/kivy/kivy/blob/c492a33f8cf79e89ea7240690feb5f6d25b08389/kivy/config.py#L902-L908

The best thing to do would be to somehow *call* the code in the above kivy/config.py script, but I'm not sure how to do it because its not in any method; it's just at the root of the script. I created this question on SE about it here:

 * https://stackoverflow.com/questions/78216476/how-to-get-kivy-to-set-its-defaults-in-the-config-at-runtime

See also:

 * https://github.com/BusKill/buskill-app/issues/55#issuecomment-2016994777