fixed value_human not being updated
This commit fixes the bug where the Setting did not list the current value if the Setting used value_human.

I updated the enable_option() function to update 'value_human' (in addition to just 'value') where 'value' is a list.

Previously the font name was being updated on the ComplexOptionScreen, but not hte main Settings screen. This is now fixed.

We do still have bug where the same field does not get updated when the reset button is clicked

TODO: fix the BusKillSettingItem's labellayout so that it does in-fact get updated after the reset button is pressed.