fixed another "radio button icon not being" set bug
This commit fixes another issue where the font that was set in the Config wasn't actually given a "checked"radio button in the GUI's ComplexOptions Screen for the Setting.

The root of the problem was our recent change from storing gui_font_face as a single string into a list as default_font. Actually, in some instances the default_font is a list and in other cases it's a string representation of a list. So the problem was that the "==" conditional was failing because one was type "list" and one was type "string"

	print( "testing if |" +str(type(BusKillApp.manager.current_screen.rv.data[n]['value']))+ "| == |" +str(type(self.value))+ "| == |"+ str(type(self.parent_option.value))+ "| )

	output:

		testing if |<class 'list'>| == |<class 'list'>| == |<class 'str'>|

The solution was to just wrap all of the variables in str() before the conditional check, and now it works!