update all Labels with new font as soon as it's selected
This commit adds a loop not just to refresh_values() but also enable_option() to loop through literally every widget in every screen and update the font.

Now the changes are applied as soon as the user clicks on the font.

Also this fixes a bug where labels on other screens would still be rendered with an old font, including dialog boxes.