fixed resetting of radio icon label on reset settings
This commit fixed a bug where, after a user clicked the "reset" button on the Settings Screen, the Config would be updated but the radio button for the previous value would still have the "checked" icon on the BusKillComplexOptionsScreen.

I fixed this by iterating through the RecycleView data dictionary in BusKillSettingsScreen.refresh_values(). Simply looping through the widgets does not work now that we've switched to a RecycleView, and many of those widgets are cached out of existence; we must update the actual data field of the rv.

There's also a bug where "Roboto" isn't selected after reset, but I think I'm going to just leave it as-is. The problem is the diff between the kivy 'default_font' and the font that we populate (and then deduplicate) with our list of fonts

	['Roboto', 'data/fonts/Roboto-Regular.ttf', 'data/fonts/Roboto-Italic.ttf', 'data/fonts/Roboto-Bold.ttf', 'data/fonts/Roboto-BoldItalic.ttf']|

	...vs...

	['Roboto-Regular', '/tmp/kivy_appdir/opt/python3.7/lib/python3.7/site-packages/kivy/data/fonts/Roboto-Regular.ttf', '/tmp/kivy_appdir/opt/python3.7/lib/python3.7/site-packages/kivy/data/fonts/Roboto-Regular.ttf', '/tmp/kivy_appdir/opt/python3.7/lib/python3.7/site-packages/kivy/data/fonts/Roboto-Regular.ttf']

It's non-trivial and not important to fix, IMHO.