fixed another bug where the radio buttons wouldn't update from the default 'U' as you scrolled down the list of fonts
This commit changes the loops where rv.data is updated for the RecycleView. I don't know why I had that additional if statement there, but removing it fixes the bug and doesn't make the app any less performant.

Before this commit if you opened the app fresh, Clicked "Settings", clicked "Font Face", and then began to scroll-down, you'd quickly encounter OptionItems where the radio button icon was neither an unchecked radio button nor a checked radio button -- it was my "U" default when instantiating the OptionItems. This meant an issue on init2()

Despite lots of careful attempts and debugging, I'm still not able to update the Dialog Labels text. It does change on rebooting the app, though. I think I'm just going to ignore this minor bug for now