fixed logic bug on radio button icons
This commit changes the conditional that was checking if the wrong thing was equal before marking the OptionItem's radio button icon as "checked" or not.

I know now what the if statement I removed in the last commit was: it made sure that I only updated the one item in the rv.data dict for that one OptionItem, because its within init2().

I would expect that init2() is going to be called once for each OptionItem, so not having that if means that every OptionItem is going to be updating every other OptionItem, which is like an unnecessary loop over a loop. Alas, it doesn't harm the app (because of the way RecycleViews work that first loop doesn't complete) as it appears to run without slowing things down, and it *does* harm the app if I leave that logic in-place, so I guess I'll leave it