combined the rv.data updates into one function
This commit adds a function ComplexOptionsScreen.update_data(), which updates all of the widgets on the ComplexOptionsScreen via its Recycle View's list of dictionaries, which is used as the backend for the widgets that are created & destroyed as-needed in & out of the cache

This saves a lot of redundant code, which I'll cleanup in the next commit.

Note we also added a new instance field 'screen' to the OptionItem because, for some reason, init2() was being called on OptionItem even when BusKillApp.manager.current_screen was *not* a ComplexOptionsScreen. So now any time an OptionItem is instantiated, we can directly reference the ComplexOptionScreen where it (and its Recycle View) lives, instead of just guessing with the 'current_screen'