replaced duplicate code with update_font_recursive() function
This commit eliminates the copied & pasted code that updated all of the fonts in two places. Now we just have one function that's called in both places

Note: This code still doesn't update any of the Label's fonts in the dialog modals at runtime. I confirmed that we are iterating over them, and even the non-instantiated ones don't take the new default_font. I can't explain this, but I'll consider it a minor bug and ignore it. Restart the app does make the dialogs' font change.